fix(testing): move cypress tsconfig to cypress folder
