+++ packages/cypress/src/generators/cypress-component-project/cypress-component-project.ts
@@ -40,6 +40,8 @@ export async function cypressComponentProject(
   addProjectFiles(tree, projectConfig, options);
   addTargetToProject(tree, projectConfig, options);
   updateNxJson(tree);
+  updateTsConfigForComponentTesting(tree, projectConfig);
+
   if (!options.skipFormat) {
     await formatFiles(tree);
   }
@@ -131,4 +133,62 @@ function updateNxJson(tree: Tree) {
     ];
   }
   updateWorkspaceConfiguration(tree, workspaceConfiguration);
+}
+
+export function updateTsConfigForComponentTesting(
+  tree: Tree,
+  projectConfig: ProjectConfiguration
+) {
+  const tsConfigPath = joinPathFragments(
+    projectConfig.root,
+    projectConfig.projectType === 'library'
+      ? 'tsconfig.lib.json'
+      : 'tsconfig.app.json'
+  );
+  if (tree.exists(tsConfigPath)) {
+    updateJson(tree, tsConfigPath, (json) => {
+      const excluded = new Set([
+        ...(json.exclude || []),
+        'cypress/**/*',
+        'cypress.config.ts',
+        '**/*.cy.ts',
+        '**/*.cy.js',
+        '**/*.cy.tsx',
+        '**/*.cy.jsx',
+      ]);
+
+      json.exclude = Array.from(excluded);
+      return json;
+    });
+  }
+
+  const projectBaseTsConfig = joinPathFragments(
+    projectConfig.root,
+    'tsconfig.json'
+  );
+  if (tree.exists(projectBaseTsConfig)) {
+    updateJson(tree, projectBaseTsConfig, (json) => {
+      if (json.references) {
+        const hasCyTsConfig = json.references.some(
+          (r) => r.path === './cypress/tsconfig.cy.json'
+        );
+        if (!hasCyTsConfig) {
+          json.references.push({ path: './cypress/tsconfig.cy.json' });
+        }
+      } else {
+        const excluded = new Set([
+          ...(json.exclude || []),
+          'cypress/**/*',
+          'cypress.config.ts',
+          '**/*.cy.ts',
+          '**/*.cy.js',
+          '**/*.cy.tsx',
+          '**/*.cy.jsx',
+        ]);
+
+        json.exclude = Array.from(excluded);
+      }
+      return json;
+    });
+  }
 }
