+++ packages/react/src/generators/cypress-component-configuration/lib/update-configs.ts
@@ -9,64 +9,6 @@ import {
 } from '@nrwl/devkit';
 import { CypressComponentConfigurationSchema } from '../schema';
 
-export function updateTsConfig(
-  tree: Tree,
-  projectConfig: ProjectConfiguration
-) {
-  const tsConfigPath = joinPathFragments(
-    projectConfig.root,
-    projectConfig.projectType === 'library'
-      ? 'tsconfig.lib.json'
-      : 'tsconfig.app.json'
-  );
-  if (tree.exists(tsConfigPath)) {
-    updateJson(tree, tsConfigPath, (json) => {
-      const excluded = new Set([
-        ...(json.exclude || []),
-        'cypress/**/*',
-        'cypress.config.ts',
-        '**/*.cy.ts',
-        '**/*.cy.js',
-        '**/*.cy.tsx',
-        '**/*.cy.jsx',
-      ]);
-
-      json.exclude = Array.from(excluded);
-      return json;
-    });
-  }
-
-  const projectBaseTsConfig = joinPathFragments(
-    projectConfig.root,
-    'tsconfig.json'
-  );
-  if (tree.exists(projectBaseTsConfig)) {
-    updateJson(tree, projectBaseTsConfig, (json) => {
-      if (json.references) {
-        const hasCyTsConfig = json.references.some(
-          (r) => r.path === './tsconfig.cy.json'
-        );
-        if (!hasCyTsConfig) {
-          json.references.push({ path: './tsconfig.cy.json' });
-        }
-      } else {
-        const excluded = new Set([
-          ...(json.exclude || []),
-          'cypress/**/*',
-          'cypress.config.ts',
-          '**/*.cy.ts',
-          '**/*.cy.js',
-          '**/*.cy.tsx',
-          '**/*.cy.jsx',
-        ]);
-
-        json.exclude = Array.from(excluded);
-      }
-      return json;
-    });
-  }
-}
-
 export async function updateProjectConfig(
   tree: Tree,
   options: CypressComponentConfigurationSchema
